+++ MediaBrowser.Controller/Persistence/IChapterRepository.cs
@@ -1,12 +1,11 @@
 using System;
 using System.Collections.Generic;
-using MediaBrowser.Model.Dto;
 using MediaBrowser.Model.Entities;
 
-namespace MediaBrowser.Controller.Chapters;
+namespace MediaBrowser.Controller.Persistence;
 
 /// <summary>
-/// Interface IChapterManager.
+/// Interface IChapterRepository.
 /// </summary>
 public interface IChapterRepository
 {
@@ -16,21 +15,6 @@ public interface IChapterRepository
     /// <param name="itemId">The item.</param>
     /// <param name="chapters">The set of chapters.</param>
     void SaveChapters(Guid itemId, IReadOnlyList<ChapterInfo> chapters);
-
-    /// <summary>
-    /// Gets all chapters associated with the baseItem.
-    /// </summary>
-    /// <param name="baseItem">The baseitem.</param>
-    /// <returns>A readonly list of chapter instances.</returns>
-    IReadOnlyList<ChapterInfo> GetChapters(BaseItemDto baseItem);
-
-    /// <summary>
-    /// Gets a single chapter of a BaseItem on a specific index.
-    /// </summary>
-    /// <param name="baseItem">The baseitem.</param>
-    /// <param name="index">The index of that chapter.</param>
-    /// <returns>A chapter instance.</returns>
-    ChapterInfo? GetChapter(BaseItemDto baseItem, int index);
 
     /// <summary>
     /// Gets all chapters associated with the baseItem.
