+++ MediaBrowser.Providers/MediaInfo/ProbeProvider.cs
@@ -55,13 +55,11 @@ namespace MediaBrowser.Providers.MediaInfo
         /// </summary>
         /// <param name="mediaSourceManager">Instance of the <see cref="IMediaSourceManager"/> interface.</param>
         /// <param name="mediaEncoder">Instance of the <see cref="IMediaEncoder"/> interface.</param>
-        /// <param name="itemRepo">Instance of the <see cref="IItemRepository"/> interface.</param>
         /// <param name="blurayExaminer">Instance of the <see cref="IBlurayExaminer"/> interface.</param>
         /// <param name="localization">Instance of the <see cref="ILocalizationManager"/> interface.</param>
-        /// <param name="encodingManager">Instance of the <see cref="IEncodingManager"/> interface.</param>
+        /// <param name="chapterManager">Instance of the <see cref="IChapterManager"/> interface.</param>
         /// <param name="config">Instance of the <see cref="IServerConfigurationManager"/> interface.</param>
         /// <param name="subtitleManager">Instance of the <see cref="ISubtitleManager"/> interface.</param>
-        /// <param name="chapterManager">Instance of the <see cref="IChapterRepository"/> interface.</param>
         /// <param name="libraryManager">Instance of the <see cref="ILibraryManager"/> interface.</param>
         /// <param name="loggerFactory">Instance of the <see cref="ILoggerFactory"/>.</param>
         /// <param name="fileSystem">Instance of the <see cref="IFileSystem"/> interface.</param>
@@ -72,13 +70,11 @@ namespace MediaBrowser.Providers.MediaInfo
         public ProbeProvider(
             IMediaSourceManager mediaSourceManager,
             IMediaEncoder mediaEncoder,
-            IItemRepository itemRepo,
             IBlurayExaminer blurayExaminer,
             ILocalizationManager localization,
-            IEncodingManager encodingManager,
+            IChapterManager chapterManager,
             IServerConfigurationManager config,
             ISubtitleManager subtitleManager,
-            IChapterRepository chapterManager,
             ILibraryManager libraryManager,
             IFileSystem fileSystem,
             ILoggerFactory loggerFactory,
@@ -96,13 +92,11 @@ namespace MediaBrowser.Providers.MediaInfo
                 loggerFactory.CreateLogger<FFProbeVideoInfo>(),
                 mediaSourceManager,
                 mediaEncoder,
-                itemRepo,
                 blurayExaminer,
                 localization,
-                encodingManager,
+                chapterManager,
                 config,
                 subtitleManager,
-                chapterManager,
                 libraryManager,
                 _audioResolver,
                 _subtitleResolver,
@@ -113,7 +107,6 @@ namespace MediaBrowser.Providers.MediaInfo
                 loggerFactory.CreateLogger<AudioFileProber>(),
                 mediaSourceManager,
                 mediaEncoder,
-                itemRepo,
                 libraryManager,
                 _lyricResolver,
                 lyricManager,
